Factored the calculation of a day's results out of the server
I'll reuse this in the CLI so that they calculate results in the
same way.
